% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_group_by_sanddance.R
\name{aes_d}
\alias{aes_d}
\title{Aesthetic mappings for datasets
Some sourcing, but not exact: https://github.com/liamgilbey/ggwaffle/blob/master/R/aes_d.R}
\usage{
aes_d(...)
}
\arguments{
\item{...}{Unquoted, comma-seperated column mappings}
}
\description{
The idea comes straight from \code{aes} in ggplot2. That provides a way to map columns of a dataset to features of graphic.
Here we are mapping columns into a function so we can use standard names inside that function.
}
\details{
A mapping looks like: <column_to_be_created> = <existing column>
}
\examples{
\dontrun{
aes_d(group = class)
}
}
