% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_group_by_sanddance.R
\name{animate_group_by_sanddance}
\alias{animate_group_by_sanddance}
\title{Produces frames (plots) of transition from ungrouped to grouped icon arrays}
\usage{
animate_group_by_sanddance(
  .data,
  ...,
  nframes = 5,
  is_last = FALSE,
  titles = ""
)
}
\arguments{
\item{.data}{the grouped dataframe TODO: or coords??????}

\item{...}{the grouping variables}
}
\value{
ggplot object
}
\description{
Produces frames (plots) of transition from ungrouped to grouped icon arrays
}
