% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_limits_list}
\alias{build_limits_list}
\title{Build list of limits (for later use of tweening between)}
\usage{
build_limits_list(xlims, ylims, id_name)
}
\arguments{
\item{xlims}{a vector of x limits}

\item{ylims}{a vector of y limits}

\item{id_name}{}
}
\value{
A list of limits, one element for every pair of limits in \code{xlims} and \code{ylims}. Each element is a data frame containing \code{xlim}, \code{ylim}, \code{time} (the same as the element number) and an ID, named as specified in \code{id_name}, that describes if each set of values is the minimum (1) or maximum (2) limit value.
}
\description{
Build list of limits (for later use of tweening between)
}
