% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamation_sanddance.R
\name{datamation_sanddance}
\alias{datamation_sanddance}
\title{Generate a plot datamation}
\usage{
datamation_sanddance(
  pipeline,
  envir = rlang::global_env(),
  pretty = TRUE,
  elementId = NULL,
  height = 300,
  width = 300
)
}
\arguments{
\item{pipeline}{Tidyverse pipeline}

\item{envir}{Environment where code is evaluated. Defaults to the global environment.}

\item{pretty}{Whether to pretty the JSON output. Defaults to TRUE.}

\item{elementId}{Optional ID for the widget element.}

\item{height}{Height of the plotting area of the widget (excluding axes and legends). This is an approximation and not an exact science, since sizes may vary depending on labels, legends, facets, etc! Defaults to 300 (pixels).}

\item{width}{Width of the plotting area of the widget (excluding axes and legends). This is an approximation and not an exact science, since sizes may vary depending on labels, legends, facets, etc! Defaults to 300 (pixels).}
}
\description{
Generates the vega lite specs necessary for a plot datamation, based on a tidyverse pipeline.
}
