% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamationSandDance.R
\name{datamationSandDance}
\alias{datamationSandDance}
\title{datamations SandDance html widget}
\usage{
datamationSandDance(specs, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{specs}{JSON specs for pipeline}

\item{width}{Width of widget. Not currently used}

\item{height}{Height of widget. Not currently used}

\item{elementId}{Optional element ID for widget}
}
\description{
datamations SandDance html widget
}
