% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_summarize_mean_sanddance.R
\name{iron_groups}
\alias{iron_groups}
\title{Helper fn to update df colnames with mapping information
Similar to waffle_iron_*}
\usage{
iron_groups(.data, mapping, geom)
}
\arguments{
\item{mapping}{created by aes_d}
}
\description{
Helper fn to update df colnames with mapping information
Similar to waffle_iron_*
}
