% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jeter_justice}
\alias{jeter_justice}
\title{Hit data for Derek Jeter and David Justice from the 1995 and 1996 seasons}
\format{
An object of class \code{data.frame} with 1181 rows and 3 columns.
}
\usage{
jeter_justice
}
\description{
Hit data for Derek Jeter and David Justice from the 1995 and 1996 seasons
}
\keyword{datasets}
