% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pad_limits}
\alias{pad_limits}
\title{Pad limits from a plot}
\usage{
pad_limits(plot)
}
\arguments{
\item{plot}{Input plot to take limits from}
}
\value{

}
\description{
Pad limits from a plot by making both the x and y limits larger in absolute size.
}
