% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_functions.R
\name{parse_functions}
\alias{parse_functions}
\title{Parse functions from pipeline steps}
\usage{
parse_functions(fittings)
}
\arguments{
\item{fittings}{List of pipeline steps}
}
\description{
Parse functions from pipeline steps
}
