% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pipeline.R
\name{parse_pipeline}
\alias{parse_pipeline}
\title{Parse a tidyverse pipeline}
\usage{
parse_pipeline(pipeline, supported_tidy_functions = c("group_by", "summarize"))
}
\arguments{
\item{pipeline}{Input pipeline, as a string.}

\item{supported_tidy_functions}{Functions that are supported by datamations: \code{group_by} and \code{summarize}/\code{summarise}.}
}
\description{
Parses a tidyverse pipeline, input as a string, into a list of its components as expressions for parsing later on in the datamations process.
}
\examples{
"small_salary \%>\% group_by(Degree) \%>\% summarize(mean = mean(Salary))" \%>\%
  datamations:::parse_pipeline()

"group_by(small_salary, Degree) \%>\% summarize(mean = mean(Salary))" \%>\%
  datamations:::parse_pipeline()
}
