% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grouped_dataframe_sanddance_.R
\name{plot_grouped_dataframe_sanddance}
\alias{plot_grouped_dataframe_sanddance}
\title{Plot grouped dataframe, in sanddance style (icon array). Should look like a bunch of icon
arrays spaced out along x axis.}
\usage{
plot_grouped_dataframe_sanddance(
  coords,
  xlim = NULL,
  ylim = NULL,
  is_coord_equal = TRUE,
  mapping = NULL,
  in_flight = FALSE,
  title = ""
)
}
\arguments{
\item{coords}{with x, y, plus optional aesthetics columns}
}
\value{
ggplot object
}
\description{
Plot grouped dataframe, in sanddance style (icon array). Should look like a bunch of icon
arrays spaced out along x axis.
}
