% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grouped_dataframe_withresponse_sanddance.R
\name{plot_grouped_dataframe_withresponse_sanddance}
\alias{plot_grouped_dataframe_withresponse_sanddance}
\title{plot a dataframe with grouping, but instead of icon array, do a jittered plot
inferring the response variable to keep interface clean???}
\usage{
plot_grouped_dataframe_withresponse_sanddance(
  .data,
  response_var,
  xlim = NULL,
  ylim = NULL,
  mapping = NULL,
  var_levels = NULL,
  in_flight = TRUE,
  show_y_axis = FALSE,
  title = ""
)
}
\arguments{
\item{response_var}{class name???}
}
\description{
plot a dataframe with grouping, but instead of icon array, do a jittered plot
inferring the response variable to keep interface clean???
}
