% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_specs_group_by.R
\name{prep_specs_group_by}
\alias{prep_specs_group_by}
\title{Produces frames (plots) of transition from ungrouped to grouped icon arrays}
\usage{
prep_specs_group_by(.data, ..., pretty = TRUE)
}
\arguments{
\item{.data}{grouped dataframe}

\item{...}{grouping variables}

\item{nframes}{number of frames per animation stage}

\item{is_last}{flag for whether this is the last stage of the pipeline. Defaults to \code{FALSE}}

\item{titles}{titles for animation stage}
}
\value{
ggplot object
}
\description{
Produces frames (plots) of transition from ungrouped to grouped icon arrays
}
