% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_specs_summarize.R
\name{prep_specs_summarize}
\alias{prep_specs_summarize}
\title{Generate specs of data distribution and summarized data}
\usage{
prep_specs_summarize(
  .data,
  summary_operation,
  toJSON = TRUE,
  pretty = TRUE,
  height = 300,
  width = 300
)
}
\arguments{
\item{.data}{Optionally grouped input data}

\item{summary_operation}{Summary operation including summary function and column operated on.}

\item{toJSON}{Whether to converts the spec to JSON. Defaults to TRUE.}

\item{pretty}{Whether to pretty the JSON output. Defaults to TRUE.}

\item{height}{Height of the plotting area of the widget (excluding axes and legends). This is an approximation and not an exact science, since sizes may vary depending on labels, legends, facets, etc! Defaults to 300 (pixels).}

\item{width}{Width of the plotting area of the widget (excluding axes and legends). This is an approximation and not an exact science, since sizes may vary depending on labels, legends, facets, etc! Defaults to 300 (pixels).}
}
\description{
Generate specs of data distribution and summarized data
}
