% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{salary_data}
\alias{salary_data}
\title{Hypothetical salary data for 3000 workers}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3000 rows and 4 columns.
}
\usage{
salary_data
}
\description{
Hypothetical salary data for 3000 workers
}
\keyword{datasets}
