% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_salary}
\alias{small_salary}
\title{Hypothetical salary data for 30 workers}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 100 rows and 3 columns.
}
\usage{
small_salary
}
\description{
Hypothetical salary data for 30 workers
}
\keyword{datasets}
