% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_salary_data}
\alias{small_salary_data}
\title{Hypothetical salary data for 30 workers}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 30 rows and 3 columns.
}
\usage{
small_salary_data
}
\description{
Hypothetical salary data for 30 workers
}
\keyword{datasets}
