% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snake.R
\name{snake}
\alias{snake}
\title{Evaluate each pipeline step}
\usage{
snake(fittings, envir = parent.frame())
}
\arguments{
\item{fittings}{Pipeline steps}

\item{envir}{Evaluation environment}
}
\description{
Evaluate each pipeline step
}
