% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-inflight.R
\name{theme_inflight}
\alias{theme_inflight}
\title{minimal theme that only shows legend and x axis ticks}
\usage{
theme_inflight(show_y_axis = FALSE)
}
\description{
minimal theme that only shows legend and x axis ticks
}
