% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_group_by_sanddance.R
\name{waffle_iron_groups}
\alias{waffle_iron_groups}
\title{Calculates the x,y, and other aesthetics}
\usage{
waffle_iron_groups(data, mapping, rows = 7, sample_size = 1, na.rm = T)
}
\arguments{
\item{data}{the dataframe to be visualized}

\item{mapping}{includes group and x, z.B.}
}
\value{
dataframe with x, y, plus optional aesthetics columns
}
\description{
Calculates the x,y, and other aesthetics
}
