% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamation_sanddance.R
\name{datamation_sanddance}
\alias{datamation_sanddance}
\title{Create a tibble datamation}
\usage{
datamation_sanddance(pipeline, envir = rlang::global_env(), pretty = TRUE)
}
\arguments{
\item{pipeline}{A tidyverse pipeline.}

\item{envir}{An environment.}

\item{output}{Path to where gif will be saved.}
}
\description{
Create a tibble datamation
}
