% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamation_sanddance.R
\name{datamation_sanddance}
\alias{datamation_sanddance}
\title{Generate a plot datamation}
\usage{
datamation_sanddance(pipeline, envir = rlang::global_env(), pretty = TRUE)
}
\arguments{
\item{pipeline}{Tidyverse pipeline}

\item{envir}{Environment where code is evaluated. Defaults to the global environment.}

\item{pretty}{Whether to pretty the JSON output. Defaults to TRUE.}
}
\description{
Generates the vega lite specs necessary for a plot datamation, based on a tidyverse pipeline.
}
