% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamationSandDance.R
\name{datamationSandDance}
\alias{datamationSandDance}
\title{datamations SandDance html widget}
\usage{
datamationSandDance(specs, width = NULL, height = NULL, elementId = NULL)
}
\description{
datamations SandDance html widget
}
