% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_specs_data.R
\name{prep_specs_data}
\alias{prep_specs_data}
\title{Generate spec of data in ungrouped icon array}
\usage{
prep_specs_data(.data, ..., toJSON = TRUE, pretty = TRUE)
}
\arguments{
\item{.data}{Input data}

\item{...}{Additional arguments, unused.}

\item{toJSON}{Whether to converts the spec to JSON. Defaults to TRUE.}

\item{pretty}{Whether to pretty the JSON output of the spec. Defaults to TRUE, and only relevant when \code{toJSON} is TRUE.}
}
\description{
Generate spec of data in ungrouped icon array
}
