% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_specs_data.R
\name{prep_specs_data}
\alias{prep_specs_data}
\title{Generate spec of data in ungrouped icon array}
\usage{
prep_specs_data(
  .data,
  mapping,
  toJSON = TRUE,
  pretty = TRUE,
  height = 300,
  width = 300
)
}
\arguments{
\item{.data}{Input data}

\item{mapping}{Mapping, unused.}

\item{toJSON}{Whether to converts the spec to JSON. Defaults to TRUE.}

\item{pretty}{Whether to pretty the JSON output. Defaults to TRUE.}

\item{height}{Height of the plotting area of the widget (excluding axes and legends). This is an approximation and not an exact science, since sizes may vary depending on labels, legends, facets, etc! Defaults to 300 (pixels).}

\item{width}{Width of the plotting area of the widget (excluding axes and legends). This is an approximation and not an exact science, since sizes may vary depending on labels, legends, facets, etc! Defaults to 300 (pixels).}
}
\description{
Generate spec of data in ungrouped icon array
}
