% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_specs_group_by.R
\name{prep_specs_group_by}
\alias{prep_specs_group_by}
\title{Generate specs of data in grouped icon array}
\usage{
prep_specs_group_by(.data, ..., pretty = TRUE)
}
\arguments{
\item{.data}{Input data}

\item{...}{Grouping variables}

\item{pretty}{Whether to pretty the JSON output of the spec. Defaults to TRUE.}
}
\description{
Generate specs of data in grouped icon array
}
